# public/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Nexus CRM API",
    "version": "1.0.0",
    "description": "REST API for the Nexus CRM application.\n\n## Authentication\n\nTwo auth methods are supported:\n\n- **BearerAuth** \u2013 `Authorization: Bearer <token>`. Per-user API token generated in the dashboard. Scoped to the token owner's data (admin tokens get cross-tenant read access).\n- **CookieAuth** \u2013 Session cookie set by the Better Auth `/api/auth` flow (Google OAuth).\n\nToken management (POST/DELETE `/api/token`) requires a session cookie. File downloads require CookieAuth or BearerAuth; unauthenticated sharing uses per-link `/s/{code}` URLs scoped to the link owner."
  },
  "servers": [
    {
      "url": "/",
      "description": "Current host"
    }
  ],
  "components": {
    "securitySchemes": {
      "BearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "description": "Per-user API token generated via the dashboard (POST /api/token). Scoped to the token owner's data. Admin users' tokens get cross-tenant read access."
      },
      "CookieAuth": {
        "type": "apiKey",
        "in": "cookie",
        "name": "better-auth.session_token",
        "description": "Session cookie issued by the Better Auth flow at `/api/auth`. Admin users (isAdmin=true) bypass per-user filtering."
      }
    },
    "schemas": {
      "ApplicationStatus": {
        "type": "string",
        "enum": ["inbound", "applied", "interview", "offer", "rejected"],
        "description": "Current stage of the opportunity pipeline."
      },
      "ContactRef": {
        "type": "object",
        "description": "A contact person associated with an application.",
        "required": ["id", "name", "applicationId", "createdAt"],
        "properties": {
          "id":            { "type": "integer", "example": 1 },
          "name":          { "type": "string",  "example": "Jane Smith" },
          "email":         { "type": ["string", "null"], "format": "email", "example": "jane@example.com" },
          "phone":         { "type": ["string", "null"], "example": "+1-555-0100" },
          "role":          { "type": ["string", "null"], "example": "Engineering Manager" },
          "linkedIn":      { "type": ["string", "null"], "example": "https://linkedin.com/in/janesmith" },
          "applicationId": { "type": "integer", "example": 42 },
          "createdAt":     { "type": "string",  "format": "date-time" }
        }
      },
...
      "Application": {
        "type": "object",
        "description": "An opportunity record.",
        "required": ["id", "userId", "company", "role", "status", "createdAt", "updatedAt", "contacts"],
        "properties": {
          "id":             { "type": "integer",              "example": 42 },
          "userId":         { "type": "string",               "example": "user_abc123" },
          "company":        { "type": "string",               "example": "Acme Corp" },
          "role":           { "type": "string",               "example": "Senior Engineer" },
          "status":         { "$ref": "#/components/schemas/ApplicationStatus" },
          "appliedAt":      { "type": ["string", "null"],     "format": "date-time" },
          "lastContact":    { "type": ["string", "null"],     "format": "date-time" },
          "followUpAt":     { "type": ["string", "null"],     "format": "date-time" },
          "notes":          { "type": ["string", "null"],     "example": "Had a great first call." },
          "jobDescription": { "type": ["string", "null"],     "example": "We are looking for…" },
          "canonicalJobUrl": { "type": ["string", "null"], "format": "uri" },
          "workMode": { "type": ["string", "null"], "enum": ["remote", "hybrid", "onsite", "flexible", null] },
          "eligibleCountries": { "type": "array", "items": { "type": "string" } },
          "primaryLocations": { "type": "array", "items": { "type": "string" } },
          "officeDaysMin": { "type": ["integer", "null"] },
          "travelPercent": { "type": ["integer", "null"] },
          "visaSponsorship": { "type": ["boolean", "null"] },
          "rightToWorkRequired": { "type": ["boolean", "null"] },
          "timezoneOverlap": { "type": ["string", "null"] },
          "salaryCurrency": { "type": ["string", "null"] },
          "salaryPeriod": { "type": ["string", "null"] },
          "salaryType": { "type": ["string", "null"] },
          "atsName": { "type": ["string", "null"] },
          "requisitionId": { "type": ["string", "null"] },
          "jobCapturedAt": { "type": ["string", "null"], "format": "date-time" },
          "jobVerifiedAt": { "type": ["string", "null"], "format": "date-time" },
          "jobPostedAt": { "type": ["string", "null"], "format": "date-time" },
          "jobClosedAt": { "type": ["string", "null"], "format": "date-time" },
          "jobContentHash": { "type": ["string", "null"] },
          "jobLiveness": { "type": ["string", "null"] },
          "jobSummary": { "type": ["string", "null"] },
          "currentStage": { "type": ["string", "null"] },
          "createdAt":      { "type": "string",               "format": "date-time" },
          "updatedAt":      { "type": "string",               "format": "date-time" },
          "contacts":       {
            "type": "array",
            "items": { "$ref": "#/components/schemas/ContactRef" }
          }
        }
      },
...
      "SubmissionAnswer": {
        "type": "object",
        "required": ["question", "answer"],
        "properties": {
          "key": { "type": ["string", "null"] },
          "question": { "type": "string", "maxLength": 1000 },
          "answer": { "type": "string", "maxLength": 20000 },
          "kind": { "type": ["string", "null"] },
          "sensitive": { "type": "boolean", "default": false }
        }
      },
      "ApplicationSubmission": {
        "type": "object",
        "description": "Immutable snapshot of one application submission. Exact answers are sensitive.",
        "required": ["id", "applicationId", "idempotencyKey", "requestHash", "submittedAt", "documentIds"],
        "properties": {
          "id": { "type": "string" },
          "applicationId": { "type": "string" },
          "idempotencyKey": { "type": "string" },
          "requestHash": { "type": "string" },
          "submittedAt": { "type": "string", "format": "date-time" },
          "applicationUrl": { "type": ["string", "null"] },
          "atsName": { "type": ["string", "null"] },
          "requisitionId": { "type": ["string", "null"] },
          "language": { "type": ["string", "null"] },
          "answers": { "type": "array", "items": { "$ref": "#/components/schemas/SubmissionAnswer" } },
          "candidateSalaryMin": { "type": ["number", "null"] },
          "candidateSalaryMax": { "type": ["number", "null"] },
          "candidateSalaryCurrency": { "type": ["string", "null"] },
          "candidateSalaryPeriod": { "type": ["string", "null"] },
          "candidateSalaryType": { "type": ["string", "null"] },
          "candidateSalaryFlexible": { "type": "boolean" },
          "documentIds": { "type": "array", "items": { "type": "string" } }
        }
      },
...
      "ErrorResponse": {
        "type": "object",
        "required": ["error"],
        "properties": {
          "error": { "type": "string", "example": "Unauthorized" }
        }
      },
...
    },
    "responses": {
      "Unauthorized": {
        "description": "Missing or invalid credentials.",
        "content": {
          "application/json": {
            "schema": { "$ref": "#/components/schemas/ErrorResponse" },
            "example": { "error": "Unauthorized" }
          }
        }
      },
      "Forbidden": {
        "description": "Insufficient permissions — admin privileges required or resource does not belong to the authenticated user.",
        "content": {
          "application/json": {
            "schema": { "$ref": "#/components/schemas/ErrorResponse" },
            "example": { "error": "Session required" }
          }
        }
      },
      "NotFound": {
        "description": "Resource not found or does not belong to the authenticated user.",
        "content": {
          "application/json": {
            "schema": { "$ref": "#/components/schemas/ErrorResponse" },
            "example": { "error": "Not found" }
          }
        }
      },
      "BadRequest": {
        "description": "Invalid or missing request parameters.",
        "content": {
          "application/json": {
            "schema": { "$ref": "#/components/schemas/ErrorResponse" }
          }
        }
      }
    }
  },
  "paths": {
    "/api/applications": {
      "get": {
        "operationId": "listApplications",
        "summary": "List applications",
        "description": "Returns all applications for the authenticated user, ordered by creation date descending. Admin users see all applications across tenants.",
        "tags": ["Applications"],
        "security": [
          { "BearerAuth": [] },
          { "CookieAuth": [] }
        ],
        "responses": {
          "200": {
            "description": "Array of applications (with embedded contacts).",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": { "$ref": "#/components/schemas/Application" }
                }
              }
            }
          },
          "401": { "$ref": "#/components/responses/Unauthorized" }
        }
      },
      "post": {
        "operationId": "createApplication",
        "summary": "Create application",
        "description": "Creates a new application owned by the session user.",
        "tags": ["Applications"],
        "security": [
          { "BearerAuth": [] },
          { "CookieAuth": [] }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": ["company", "role"],
                "properties": {
                  "company":        { "type": "string", "maxLength": 255, "example": "Acme Corp" },
                  "role":           { "type": "string", "maxLength": 255, "example": "Senior Engineer" },
                  "status":         { "$ref": "#/components/schemas/ApplicationStatus" },
                  "appliedAt":      { "type": ["string", "null"], "format": "date-time" },
                  "lastContact":    { "type": ["string", "null"], "format": "date-time" },
                  "followUpAt":     { "type": ["string", "null"], "format": "date-time" },
                  "notes":          { "type": ["string", "null"], "maxLength": 10000 },
                  "jobDescription": { "type": ["string", "null"], "maxLength": 50000 }
                }
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Newly created application.",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/Application" }
              }
            }
          },
          "400": { "$ref": "#/components/responses/BadRequest" },
          "401": { "$ref": "#/components/responses/Unauthorized" },
          "403": { "$ref": "#/components/responses/Forbidden" }
        }
      }
    },
...
    "/api/applications/{id}/submissions": {
      "parameters": [
        { "name": "id", "in": "path", "required": true, "schema": { "type": "integer", "minimum": 1 } }
      ],
      "get": {
        "operationId": "listApplicationSubmissions",
        "summary": "List immutable application submission snapshots",
        "tags": ["Applications"],
        "security": [{ "BearerAuth": [] }, { "CookieAuth": [] }],
        "parameters": [
          { "name": "includeAnswers", "in": "query", "schema": { "type": "boolean", "default": false }, "description": "Include sensitive exact answers." }
        ],
        "responses": {
          "200": {
            "description": "Submission snapshots.",
            "content": { "application/json": { "schema": { "type": "array", "items": { "$ref": "#/components/schemas/ApplicationSubmission" } } } }
          },
          "401": { "$ref": "#/components/responses/Unauthorized" },
          "404": { "$ref": "#/components/responses/NotFound" }
        }
      },
      "post": {
        "operationId": "recordApplicationSubmission",
        "summary": "Atomically record an application submission",
        "description": "Preserves exact answers, compensation, document versions, status, follow-up, and a timeline event. Supports idempotency, dry-run, and optimistic concurrency.",
        "tags": ["Applications"],
        "security": [{ "BearerAuth": [] }, { "CookieAuth": [] }],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": ["idempotencyKey", "submittedAt"],
                "properties": {
                  "idempotencyKey": { "type": "string", "minLength": 8, "maxLength": 128 },
                  "submittedAt": { "type": "string", "format": "date-time" },
                  "followUpAt": { "type": ["string", "null"], "format": "date-time" },
                  "applicationUrl": { "type": ["string", "null"], "format": "uri" },
                  "atsName": { "type": ["string", "null"] },
                  "requisitionId": { "type": ["string", "null"] },
                  "language": { "type": ["string", "null"] },
                  "answers": { "type": "array", "items": { "$ref": "#/components/schemas/SubmissionAnswer" } },
                  "candidateSalaryMin": { "type": ["number", "null"] },
                  "candidateSalaryMax": { "type": ["number", "null"] },
                  "candidateSalaryCurrency": { "type": ["string", "null"] },
                  "candidateSalaryPeriod": { "type": ["string", "null"] },
                  "candidateSalaryType": { "type": ["string", "null"] },
                  "candidateSalaryFlexible": { "type": "boolean" },
                  "documentIds": { "type": "array", "maxItems": 20, "items": { "type": "string" } },
                  "expectedUpdatedAt": { "type": "string", "format": "date-time" },
                  "dryRun": { "type": "boolean", "default": false }
                }
              }
            }
          }
        },
        "responses": {
          "201": { "description": "Recorded submission package." },
          "400": { "$ref": "#/components/responses/BadRequest" },
          "401": { "$ref": "#/components/responses/Unauthorized" },
          "409": { "description": "Idempotency or optimistic-concurrency conflict." }
        }
      }
    },
...
  },
...
}